Move 74xxGxx to a project library
The symbols in the upstream libraries use hidden power pins which
doesn't work in this design which needs different power rails on
different instances.

So far, a private modifed branch of the libraries was used. But,
becasue the upstream library won't be fix until KiCad 6 for
compatibility reasons, a project library has been created which
makes it easier to share this project.